Start fix and refactor
