+++ suspend.go
@@ -7,6 +7,8 @@ import (
 
 	// Third Party Libraries
 	"github.com/shirou/gopsutil/process"
+	// Nyrna Packages
+	// "github.com/Merrit/nyrna/find_process/find_process_linux"
 )
 
 // FindProcess is the beginning of the tree to
